RAG/requirements.txt
fitz==0.0.1.dev2
Flask==3.1.0
flask_cors==5.0.1
langchain==0.3.21
langchain_community==0.3.20
langchain_core==0.3.49
langchain_openai==0.3.11
mysql_connector_repackaged==0.3.1
Requests==2.32.3
